+++ baby_elf.py
@@ -1,0 +1,8 @@
+
+from enemy import enemy
+
+class baby_elf(enemy):
+   def take_damage(self, damage):
+      print("why are you hitting me. IM A BABY.")
+      
+      return super().take_damage(damage)
