+++ game.py
@@ -2,6 +2,9 @@ import random
 from goblin import Goblin
 from hero import Hero
 
+from witch import witch
+from baby_elf import baby_elf
+
 def main():
     print("Welcome to the Battle Arena!")
     print("༼ ᓄºل͟º ༽ᓄ   ᕦ(ò_óˇ)ᕤ")
@@ -10,7 +13,7 @@ def main():
     hero = Hero("Aragorn")
 
     # Create goblins ༼ ºل͟º ༽ ༼ ºل͟º ༽ ༼ ºل͟º ༽
-    goblins = [Goblin(f"Goblin {i+1}") for i in range(3)]
+    goblins = [Goblin(f"Goblin {i+1}", "green") for i in range(5)]
 
 
     # Keep track of how many goblins were defeated
